Move Result into transit.model.framework
