+++ src/ext/java/org/opentripplanner/ext/siri/TimetableHelper.java
@@ -18,7 +18,6 @@ import java.util.Set;
 import java.util.function.Function;
 import java.util.function.Supplier;
 import javax.xml.datatype.Duration;
-import org.opentripplanner.common.model.Result;
 import org.opentripplanner.model.PickDrop;
 import org.opentripplanner.model.StopTime;
 import org.opentripplanner.model.Timetable;
@@ -27,6 +26,7 @@ import org.opentripplanner.model.UpdateError;
 import org.opentripplanner.transit.model.basic.NonLocalizedString;
 import org.opentripplanner.transit.model.framework.Deduplicator;
 import org.opentripplanner.transit.model.framework.FeedScopedId;
+import org.opentripplanner.transit.model.framework.Result;
 import org.opentripplanner.transit.model.network.TripPattern;
 import org.opentripplanner.transit.model.site.StopLocation;
 import org.opentripplanner.transit.model.timetable.OccupancyStatus;
