+++ src/test/java/org/opentripplanner/common/model/ResultTest.java
@@ -4,6 +4,7 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+import org.opentripplanner.transit.model.framework.Result;
 
 class ResultTest {
